requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
chromadb
numpy
pydantic
python-multipart
jinja2
markdown
requests
ollama
PyPDF2
pypdf
python-latex
